Add files via upload
Implements first steps of an SVC algorithm, score at .812. 

The notebook is based on Chris' feature processing work. It scales the age feature prior to prediction (not in the original preprocessing file). 

Includes hyperparameter selection (picking the best combination of parameters for model type and regularization penalty), a validation curve for the choice of C, an ROC curve, and the confusion matrix for the model (to enable the selection of additional features).